Updated Cogs
Mine has full support for a button and is working.
Character has full button support as well as !materials and !profile command show stats. 
Items is a crucial cog that helps control items and their data.
Shop is an important cog that drives the shop end of things. It's currently in progress, and does not deliver a full shopping experience yet.